Replace gnome-keyring-daemon in .profile
This is so that, if the daemon is already running, it will be replaced, using the correct components. That is, if the daemon was started with only the secrets component (not pkcs11 or ssh), the daemon will be restarted with those components. This is necessary because I'm using gnome-keyring-daemon as my SSH key manager.